Add ThermalTransfer to PHFemEngine Fix some warnings. Add listener into PHContactDetector (but not used nor tested yet)

+++ core/src/Foundation/UTQPTimer.h
@@ -85,7 +85,7 @@ public:
 		return (int)names.size();
 	}
 	const char* GetNameOfCounter(int id) {
-		if (0<= id && id < names.size()) return names[id].name.c_str();
+		if (0<= id && id < (int)names.size()) return names[id].name.c_str();
 		return NULL;
 	}
 	const char* GetName() { return name.c_str(); }
@@ -113,7 +113,7 @@ public:
 	void SetUnit(double u) { unit = u; }
 	double GetUnit() { return unit; }
 	void ClearCounts() {
-		for (int i = 0; i < names.size(); ++i) {
+		for (int i = 0; i < (int)names.size(); ++i) {
 			counts[i] = 0;
 		}
 	}
